abstract sendHeight
