+++ packages/roadmap/src/utils/index.ts
@@ -3,3 +3,4 @@ export { default as getLinkType } from "./getLinkType";
 export { default as getLogo } from "./getLogo";
 export { default as getRoadmap } from "./getRoadmap";
 export { default as paramsToPendo } from "./paramsToPendo";
+export { default as sendHeight } from "./sendHeight";
